ticket-issue-app/admin_service/requirements.txt
flask==2.0.3
werkzeug==2.0.3
psycopg2-binary==2.9.3